# private/storage/api-docs/api-docs.json
{"openapi":"3.0.0","info":{"title":"Tribe API","description":"Tribe API Demo","version":"1.0"},"paths":{"\/oauth\/token":{"post":{"tags":["Auth"],"description":"Request an API Token resource.","operationId":"authLogin","requestBody":{"description":"Request an API Token resource.","required":true,"content":{"application\/json":{"schema":{"properties":{"grant_type":{"type":"string","example":"password"},"client_id":{"type":"integer","example":1},"client_secret":{"type":"string"},"username":{"type":"string","example":"email@sample.com"},"password":{"type":"string","example":"my-password"},"scope":{"type":"string"}}}}}},"responses":{"200":{"description":"API Token resource","content":{"application\/json":{"schema":{"properties":{"token_type":{"type":"string","example":"Bearer"},"expires_in":{"type":"integer","example":86399},"access_token":{"type":"string"},"refresh_token":{"type":"string"}}}}}},"422":{"description":"Unprocessable Entity"}}}},"\/v1\/registers":{"post":{"tags":["Auth"],"summary":"Register a new User resource.","description":"Register a new User resource","operationId":"authRegister","requestBody":{"description":"Register a new User resource","required":true,"content":{"application\/json":{"schema":{"properties":{"name":{"description":"Register a new User resource.","type":"string","example":"Name"},"email":{"description":"Register a new User resource.","type":"string","example":"user@example.com"},"password":{"description":"Register a new User resource.","type":"string","example":"my-password"},"c_password":{"description":"Register a new User resource.","type":"string","example":"my-password"}}}}}},"responses":{"200":{"description":"User collection response","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/User"}}}},"422":{"description":"Unprocessable Entity"}}}},"\/v1\/users":{"get":{"tags":["User"],"summary":"Retrieves the collection of User resources.","description":"This can only be done by the logged in user.","operationId":"indexUsers","responses":{"200":{"description":"User collection response","content":{"application\/json":{"schema":{"type":"array","items":{"$ref":"#\/components\/schemas\/User"}}}}},"401":{"description":"Unauthenticated"}},"security":[{"passport":[]}]},"post":{"tags":["User"],"summary":"Creates a User resource.","description":"This can only be done by the logged in user.","operationId":"storeUser","requestBody":{"description":"Register a new User resource","required":true,"content":{"application\/json":{"schema":{"properties":{"name":{"description":"Creates a User resource.","type":"string","example":"Name"},"email":{"description":"Creates a User resource.","type":"string","example":"user@example.com"},"password":{"description":"Creates a User resource.","type":"string","example":"my-password"}}}}}},"responses":{"201":{"description":"User resource created","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/User"}}}},"422":{"description":"Invalid input"}},"security":[{"passport":[]}]}},"\/v1\/users\/{id}":{"get":{"tags":["User"],"summary":"Retrieves a User resource.","description":"This can only be done by the logged in user.","operationId":"showUser","parameters":[{"name":"id","in":"path","description":"UUID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"User resource response","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/User"}}}},"401":{"description":"Unauthenticated"},"404":{"description":"Resource not found"}},"security":[{"passport":[]}]},"put":{"tags":["User"],"summary":"Replaces the User resource.","description":"This can only be done by the logged in user.","operationId":"updateUser","parameters":[{"name":"id","in":"path","description":"UUID","required":true,"schema":{"type":"string"}}],"requestBody":{"description":"Register a new User resource","required":true,"content":{"application\/json":{"schema":{"properties":{"name":{"description":"Replaces the User resource.","type":"string","example":"Name"},"email":{"description":"Replaces the User resource.","type":"string","example":"user@example.com"},"password":{"description":"Replaces the User resource.","type":"string","example":"my-password"}}}}}},"responses":{"201":{"description":"User resource created","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/User"}}}},"422":{"description":"Invalid input"},"404":{"description":"Resource not found"}},"security":[{"passport":[]}]},"delete":{"tags":["User"],"summary":"Removes the User resource.","description":"This can only be done by the logged in user.","operationId":"destroyUser","parameters":[{"name":"id","in":"path","description":"UUID","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"Successful delete User resource"},"401":{"description":"Unauthenticated"},"404":{"description":"Resource not found"}},"security":[{"passport":[]}]}}},"components":{"schemas":{"User":{"required":["name","email","password"],"properties":{"id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string","format":"email"},"created_at":{"type":"string","format":"date-time"},"updated_at":{"type":"string","format":"date-time"}},"type":"object"}},"securitySchemes":{"passport":{"type":"oauth2","description":"Laravel passport oauth2 security.","in":"header","scheme":"https","flows":{"password":{"authorizationUrl":"http:\/\/localhost\/oauth\/authorize","tokenUrl":"http:\/\/localhost\/oauth\/token","refreshUrl":"http:\/\/localhost\/token\/refresh","scopes":[]}}}}}}